Moved sources files into src folder
