+++ api/.dockerignore
@@ -1,13 +1,11 @@
 *
-!app.js
-!db.js
-!server.js
+!src/app.js
+!src/container.js
+!src/db.js
+!src/router.js
+!src/server.js
 !package.json
 !yarn.lock
-!components
-!controllers
-!middlewares
-!models
-!parameters
-!routes
-!services
+!src/components
+!src/middlewares
+!src/parameters
