+++ api/package.json
@@ -10,8 +10,8 @@
   },
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon --inspect=0.0.0.0:$NODE_DEBUG_PORT --nolazy --config nodemon.json server.js",
+    "start": "node ./src/server.js",
+    "dev": "nodemon --inspect=0.0.0.0:$NODE_DEBUG_PORT --nolazy --config nodemon.json ./src/server.js",
     "lint": "eslint --fix --color -f codeframe .",
     "test": "NODE_ENV=test jest --colors"
   },
