+++ api/tests/components/debug.test.js
@@ -1,4 +1,4 @@
-const debug = require('../../components/debug')('env-test');
+const debug = require('../../src/components/debug')('env-test');
 
 describe('Test debug component', () => {
   test('Should be disabled in test env', () => {
