+++ api/tests/components/logger.test.js
@@ -1,4 +1,4 @@
-const { logger } = require('../../components');
+const { logger } = require('../../src/components');
 
 describe('Test logger component', () => {
   test('Should be a function', () => {
